Pointed motion events to lastsnap.jpg
This kind of hacks over an issue where motion is not writing accurate
information to the motion-pipe file, so that file is now only being
used to see when it has been written to, triggering a new motion
event. Motion notification snapshots are now lastsnap.jpg, which
Motion is copying to from the latest motion event picture.